Moved some modules around (#115)
* made module structure more clear

removed `transformers`: everything is a transformer
moved everything from `transformers` to `preprocessing`
moved selectors in preprocessing to `selectors`
moved EstimatorTransformer to `meta`

* fix flake issue

* moved selectors back to preprocessing to avoid name clash with stdlib module

* moved selectors back to preprocessing to avoid name clash with stdlib module
